18Fall/ee838_video/hw4/source/requirements.txt
tensorflow-gpu==1.4
matplotlib
tqdm
pillow
scipy
numpy
imageio
animeface
pillow
scikit-image
